requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
google-generativeai==0.8.3
duckduckgo-search==6.3.7
httpx==0.27.2
tavily-python==0.7.23